+++ circumflex-core/src/main/scala/messages.scala
@@ -7,7 +7,7 @@ class Messages(val baseName: String, val locale: Locale) extends HashModel {
     ResourceBundle.getBundle(baseName, locale)
   } catch {
     case e => {
-      cxLog.trace("ResourceBundle for messages instance not found: " + baseName)
+      cxLog.debug("ResourceBundle for messages instance not found: " + baseName)
       null
     }
   }
@@ -16,34 +16,22 @@ class Messages(val baseName: String, val locale: Locale) extends HashModel {
   } catch {
     case e => None
   }
+  def get(key: String, params: Pair[String, String]*): Option[String] = get(key, Map(params: _*))
   def get(key: String, params: Map[String, String]): Option[String] =
-    apply(key) map {
-      params.foldLeft(_) {
-        case (m, (name, value)) => m.replaceAll("\\{" + name + "\\}", value)
-      }
-    }
-  override def apply(key: String): Option[String] = apply(key) match {
-    case None =>
+    get(key).map(m => params.foldLeft(m) {
+      case (m, (name, value)) => m.replaceAll("\\{" + name + "\\}", value)
+    })
+  override def apply(key: String): String = get(key) match {
+    case Some(v) => v
+    case _ =>
       cxLog.warn("Missing message for key {}, locale {}.", key, msgBundle.getLocale)
       ""
-    case v => v
   }
-  def apply(key: String, params: Map[String, String]): Option[String] =
-    apply(key, params) match {
-      case None =>
-        cxLog.warn("Missing message for key {}, locale {}.", key, msgBundle.getLocale)
-        ""
-      case v => v
-    }
-}
-
-object Messages {
-  def apply(): Messages = {
-    if (!CircumflexContext.live_?)
-      throw new CircumflexException("CircumflexContext is not available.")
-    ctx('msg) match {
-      case Some(m: Messages) => m
-      case _ => throw new CircumflexException("Messages instance not found in CircumflexContext.")
-    }
+  def apply(key: String, params: Pair[String, String]*): String = apply(key, Map(params: _*))
+  def apply(key: String, params: Map[String, String]): String = get(key, params) match {
+    case Some(v) => v
+    case _ =>
+      cxLog.warn("Missing message for key {}, locale {}.", key, msgBundle.getLocale)
+      ""
   }
 }